requirements.txt
autopep8==1.6.0
cffi==1.15.0
click==8.1.2
cryptography==36.0.2
Flask==2.1.1
gunicorn==20.1.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
python-dotenv==0.20.0
toml==0.10.2
Werkzeug==2.1.1
zipp==3.8.0